+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+This release adds a new annotation to document system-dependent methods, improves Representable* annotations for external processing, and fixes inherited method lookup in the Representable* processor.
+
 ### Added
 
 - Add SystemDependent annotation [#426](https://github.com/nbbrd/java-design-util/issues/426)