+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.10.0] - 2026-04-14
+
 This release adds a new annotation to document system-dependent methods, improves Representable* annotations for external processing, and fixes inherited method lookup in the Representable* processor.
 
 ### Added
@@ -145,7 +147,8 @@ This release replaces the `StringValue` annotation with a more general approach 
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-design-util/compare/v1.9.0...HEAD
+[Unreleased]: https://github.com/nbbrd/java-design-util/compare/v1.10.0...HEAD
+[1.10.0]: https://github.com/nbbrd/java-design-util/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/nbbrd/java-design-util/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/nbbrd/java-design-util/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/nbbrd/java-design-util/compare/v1.6.0...v1.7.0